Dark/light switch, zoom adaptive alpha for some sublayers
