+++ src/Theme.tsx
@@ -1,11 +1,10 @@
 import { Button } from "@radix-ui/themes";
 import { SunIcon, MoonIcon } from "@radix-ui/react-icons";
-import { useMenuStore } from "../Header.tsx";
+import { useMenuStore } from "./Store";
 
 export function ThemeButton() {
   const theme = useMenuStore((state: any) => state.theme);
   const setTheme = useMenuStore((state: any) => state.setTheme);
-  console.log(theme);
   const icon = theme == "light" ? <SunIcon /> : <MoonIcon />;
 
   return (
