+++ src/colors.module.css
@@ -1,5 +1,5 @@
 @value ground: hsla(144, 35.71%, 72.55%, 1);
-@value water: hsla(196,83.3%,92.9%,1);
+@value water: hsla(196, 83.3%, 92.9%, 1);
 @value city-dense: hsla(0, 0%, 82.35%, 1);
 @value city-sparse: hsla(0, 0%, 86.27%, 1);
 @value city-closed: hsla(0, 0%, 90.2%, 1);
