+++ src/main.tsx
@@ -1,7 +1,6 @@
 import { renderToDOM, App } from "./App.tsx";
 import { Header } from "./Header.tsx";
 import { Footer } from "./Footer.tsx";
-import { Theme } from "@radix-ui/themes";
 
 import "@radix-ui/themes/styles.css";
 import "./App.module.css";
@@ -10,29 +9,13 @@ renderToDOM(
   document.getElementById("root"),
   <div>
     <header>
-      <Theme
-        accentColor="gray"
-        grayColor="mauve"
-        panelBackground="translucent"
-        scaling="100%"
-        radius="medium"
-      >
-        <Header />
-      </Theme>
+      <Header />
     </header>
     <div id="map">
       <App />
     </div>
     <footer>
-      <Theme
-        accentColor="gray"
-        grayColor="mauve"
-        panelBackground="translucent"
-        scaling="100%"
-        radius="medium"
-      >
-        <Footer />
-      </Theme>
+      <Footer />
     </footer>
   </div>,
 );
